+++ src/Command/PostBuildTasksCommand.php
@@ -9,7 +9,7 @@ use Symfony\Component\Console\Output\OutputInterface;
 use Torr\Hosting\Deployment\TaskCli;
 use Torr\Hosting\Deployment\TaskRunners;
 
-#[AsCommand("hosting:post-build")]
+#[AsCommand("hosting:run-tasks:post-build", aliases: ["hosting:post-build"])]
 final class PostBuildTasksCommand extends Command
 {
 	private TaskRunners $runners;
@@ -27,9 +27,20 @@ final class PostBuildTasksCommand extends Command
 	/**
 	 * @inheritDoc
 	 */
-	protected function execute(InputInterface $input, OutputInterface $output) : int
+	protected function execute (InputInterface $input, OutputInterface $output) : int
 	{
 		$io = new TaskCli($input, $output);
+
+		// @todo remove in v3 and remove alias above
+		if ("hosting:run-tasks:post-build" !== $input->getFirstArgument())
+		{
+			\trigger_deprecation("21torr/hosting", "2.1.0", "The command name '{$input->getFirstArgument()}' was deprecated. Use 'hosting:run-tasks:post-build' instead.");
+			$io->caution(\sprintf(
+				"The command name `%s` was deprecated.\nUse `hosting:run-tasks:post-build` instead.",
+				$input->getFirstArgument(),
+			));
+		}
+
 		$io->title("Run Post Build Tasks");
 
 		$this->runners->runPostBuild($io);